Restructure python code
